start_plane and end_plane

+++ Source/Lib/Common/Codec/EbIntraPrediction.c
@@ -4633,8 +4633,8 @@ EbErrorType av1_intra_prediction_cl(
     PredictionMode mode;
 #if SEARCH_UV_MODE
     // Hsan: plane should be derived @ an earlier stage (e.g. @ the call of perform_fast_loop())
-    uint8_t start_plane = (md_context_ptr->uv_search_path) ? 1 : 0;
-    uint8_t end_plane = (md_context_ptr->blk_geom->has_uv && md_context_ptr->chroma_level <= CHROMA_MODE_1) ? (int)MAX_MB_PLANE : 1;
+    int32_t start_plane = (md_context_ptr->uv_search_path) ? 1 : 0;
+    int32_t end_plane = (md_context_ptr->blk_geom->has_uv && md_context_ptr->chroma_level <= CHROMA_MODE_1) ? (int)MAX_MB_PLANE : 1;
 
     for (int32_t plane = start_plane; plane < end_plane; ++plane) {
 #else
